TODGround extension for a spin scan strategy
In the context of the CMB ground-based experiments,
in particular for those that are devoted to the large
scale CMB polarization observation. They use a continuous
spin-like scanning strategy. This kind of scanning
strategy is mechanical feasible using a rotary-joint
brushless bearing.

One of them is LSPE/STRIP, a ground-based telescope,
working at 43 GHz and 95GHz for the CMB large scale
structures observation from Tenerife. For this reason,
an extension of the class `TODGround` is needed.

In this first commit, as was sugged by Reijo, I have
introduced an exception in `sim_tod.py` in order to permit
a continuos pointing generation from 0 to 2pi if, in the
schedule params, `azmin = azmax = 0`.

This exception break some part in the atmosphere
observation module `sim_det_atm.py` that now are commented,
but I have to go over in order to implement all the `if`
conditions that we are needed.